requirements.txt
fastapi
langchain-openai
langchain
python-dotenv
langchain-community 
tavily-python